Remove already-assigned match-ups from list
If a user adds 'Adelaide vs Centrals' for Round 1, remove Adelaide &
Centrals from the 'Add Match-up' select lists for Round 1 - so the user
can't add Adelaide twice.

Rounds are loaded from the database, but the database only contains
Rounds that have an associated Matchup. So fill in the games via
fillInRounds()